test/requirements.txt
# Core test/runtime packages for running the analysis unit test locally.
# Install these in your virtualenv. We intentionally exclude system-level
# components (ffmpeg, voyager index DB, etc.) which should be stubbed.
# IMPORTANT: Versions aligned with requirements/common.txt and Dockerfile
# Updated: 2026-01-08 - Aligned with Ubuntu 24.04/Python 3.12 dependencies:
#   - scipy: 1.15.3 → 1.16.3
#   - onnx: 1.16.0 → 1.20.0
#   - scikit-learn: 1.7.2 → 1.8.0
#   - transformers: unversioned → 4.57.3
#   - sentencepiece: unversioned → 0.2.1
#   - google-generativeai → google-genai==1.57.0 (package renamed)
#   - pydub: unversioned → >=0.25.1

numpy==1.26.4
librosa==0.11.0
scipy==1.16.3
numba==0.60.0
soundfile==0.13.1
resampy
onnx==1.20.0
onnxruntime==1.19.2
pydub==0.25.1
scikit-learn==1.8.0
psycopg2-binary
ftfy
pytest>=7.0.0
requests
psutil
transformers==4.57.3
sentencepiece==0.2.1

# Additional runtime dependencies needed for imports
redis
rq
Flask
pyyaml
voyager
rapidfuzz
sqlglot
six
flasgger
umap-learn
python-mpd2
google-genai==1.57.0
argon2-cffi==25.1.0
mistralai>=1.11.1,<2.0.0

# Lyrics integration test dependencies (versions aligned with requirements/common.txt).
# torch is pulled in transitively by openai-whisper. We install the CPU-only build
# from PyPI defaults; if you need a smaller wheel for local Linux/Windows installs,
# use `pip install --extra-index-url https://download.pytorch.org/whl/cpu torch==2.6.0+cpu`
# before installing this requirements file.
openai-whisper==20250625
langdetect==1.0.9
torch==2.6.0

# Notes:
# - ffmpeg (system binary) is NOT included here — keep pydub/ffmpeg calls stubbed or
#   install ffmpeg separately on your machine if you want full fallback behavior.
# - voyager is a system-level vector index; tests should keep a small stub rather
#   than installing production Voyager binaries.

